Extract EventConfigData to its own file
